refactor: components refactor
