+++ src/common/utils/categoryList.ts
@@ -1,0 +1,62 @@
+export const categoryList = [
+  {
+    name: '일반쓰레기',
+    key: 'general-waste',
+    check: false,
+  },
+  {
+    name: '플라스틱',
+    key: 'plastic',
+    check: false,
+  },
+  {
+    name: '종이류',
+    key: 'paper',
+    check: false,
+  },
+  {
+    name: '음식물쓰레기',
+    key: 'food-waste',
+    check: false,
+  },
+  {
+    name: '잡병',
+    key: 'mixed-bottle',
+    check: false,
+  },
+  {
+    name: '유리병',
+    key: 'glass-bottle',
+    check: false,
+  },
+  {
+    name: '스티로폼',
+    key: 'styrofoam',
+    check: false,
+  },
+  {
+    name: '캔류',
+    key: 'can',
+    check: false,
+  },
+  {
+    name: '헌옷수거함',
+    key: 'old-clothes-locker',
+    check: false,
+  },
+  {
+    name: '고철',
+    key: 'iron',
+    check: false,
+  },
+  {
+    name: '건전지',
+    key: 'battery',
+    check: false,
+  },
+  {
+    name: '형광등',
+    key: 'lamp',
+    check: false,
+  },
+];
