+++ src/components/BadgeList.tsx
@@ -1,12 +1,12 @@
-import {Box, HStack, Text} from 'native-base';
 import React from 'react';
+import {HStack, Text} from 'native-base';
 import styled from 'styled-components/native';
 
-type Props = {
+interface IBadgeList {
   data: Array<string>;
-};
+}
 
-const BadgeList = ({data}: Props) => {
+function BadgeList({data}: IBadgeList) {
   return (
     <HStack space={1} marginY="2">
       {data.map((item, index) => (
@@ -21,9 +21,10 @@ const BadgeList = ({data}: Props) => {
       ))}
     </HStack>
   );
-};
+}
 
 const Badge = styled(Text)`
   background-color: #d1d5db;
 `;
+
 export default BadgeList;
