+++ src/screens/Auth/LoginScreen.tsx
@@ -1,12 +1,4 @@
-import {
-  Box,
-  Button,
-  HStack,
-  Icon,
-  NativeBaseProvider,
-  Text,
-  VStack,
-} from 'native-base';
+import {Box, Button, HStack, Icon, Text, VStack} from 'native-base';
 import React from 'react';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import Input from '~/components/Input';
@@ -14,7 +6,7 @@ import {useForm, Controller} from 'react-hook-form';
 import ErrorMsg from '~/components/ErrorMsg';
 import * as yup from 'yup';
 import {yupResolver} from '@hookform/resolvers/yup';
-import {signIn} from '~/api/fireAuth';
+import {signIn} from '~/common/api/fireAuth';
 
 interface FormData {
   email: string;
@@ -42,102 +34,96 @@ function LoginScreen({navigation}: any) {
   } = useForm<FormData>(formOptions);
 
   return (
-    <NativeBaseProvider>
-      <Box h="100%" w="100%" paddingTop={20} backgroundColor="white">
-        <VStack alignItems="center" justifyContent="center">
-          <VStack space={6}>
-            <VStack space={3}>
-              <Controller
-                control={control}
-                rules={{
-                  required: true,
-                }}
-                render={({field: {onChange, onBlur, value}}) => (
-                  <Input
-                    onBlur={onBlur}
-                    onChangeText={onChange}
-                    value={value}
-                    type="text"
-                    ph="이메일을 입력해주세요"
-                  />
-                )}
-                name="email"
-              />
-              {errors.email && (
-                <ErrorMsg>이메일 형식이 올바르지 않습니다!</ErrorMsg>
-              )}
-            </VStack>
-            <VStack space={3}>
-              <Controller
-                control={control}
-                rules={{
-                  required: true,
-                }}
-                render={({field: {onChange, onBlur, value}}) => (
-                  <Input
-                    onBlur={onBlur}
-                    onChangeText={onChange}
-                    type="password"
-                    ph="비밀번호를 입력해주세요"
-                    value={value}
-                  />
-                )}
-                name="password"
-              />
-              {errors.password && (
-                <ErrorMsg>비밀번호를 다시 확인해주세요!</ErrorMsg>
-              )}
-            </VStack>
-          </VStack>
-          <VStack space={5} marginTop={16}>
-            <Button
-              colorScheme={'blue'}
-              w={96}
-              padding={'4'}
-              onPress={handleSubmit(onSubmit)}>
-              <Text color={'white'} fontSize={'16px'}>
-                로그인
-              </Text>
-            </Button>
-            <Button
-              variant="unstyled"
-              w={'96'}
-              padding={'4'}
-              onPress={() => navigation.navigate('RegisterScreen')}>
-              <Text color={'black'} fontSize={'16px'}>
-                회원가입
-              </Text>
-            </Button>
-          </VStack>
-          <VStack marginTop={16}>
-            <Button
-              backgroundColor={'light.100'}
-              w={96}
-              padding={4}
-              onPress={() => console.log('google login')}>
-              <HStack
-                w={96}
-                paddingX={4}
-                alignItems={'center'}
-                justifyContent="space-between">
-                <Icon
-                  as={Ionicons}
-                  name="logo-google"
-                  size="md"
-                  color={'black'}
+    <Box h="100%" w="100%" paddingTop={20} backgroundColor="white">
+      <VStack alignItems="center" justifyContent="center">
+        <VStack space={6}>
+          <VStack space={3}>
+            <Controller
+              control={control}
+              rules={{
+                required: true,
+              }}
+              render={({field: {onChange, onBlur, value}}) => (
+                <Input
+                  onBlur={onBlur}
+                  onChangeText={onChange}
+                  value={value}
+                  type="text"
+                  ph="이메일을 입력해주세요"
                 />
-                <Text fontSize={'16px'}>구글 아이디로 로그인</Text>
-                <Icon
-                  as={Ionicons}
-                  name="chevron-forward-outline"
-                  size={'md'}
+              )}
+              name="email"
+            />
+            {errors.email && (
+              <ErrorMsg>이메일 형식이 올바르지 않습니다!</ErrorMsg>
+            )}
+          </VStack>
+          <VStack space={3}>
+            <Controller
+              control={control}
+              rules={{
+                required: true,
+              }}
+              render={({field: {onChange, onBlur, value}}) => (
+                <Input
+                  onBlur={onBlur}
+                  onChangeText={onChange}
+                  type="password"
+                  ph="비밀번호를 입력해주세요"
+                  value={value}
                 />
-              </HStack>
-            </Button>
+              )}
+              name="password"
+            />
+            {errors.password && (
+              <ErrorMsg>비밀번호를 다시 확인해주세요!</ErrorMsg>
+            )}
           </VStack>
         </VStack>
-      </Box>
-    </NativeBaseProvider>
+        <VStack space={5} marginTop={16}>
+          <Button
+            colorScheme={'blue'}
+            w={96}
+            padding={'4'}
+            onPress={handleSubmit(onSubmit)}>
+            <Text color={'white'} fontSize={'16px'}>
+              로그인
+            </Text>
+          </Button>
+          <Button
+            variant="unstyled"
+            w={'96'}
+            padding={'4'}
+            onPress={() => navigation.navigate('RegisterScreen')}>
+            <Text color={'black'} fontSize={'16px'}>
+              회원가입
+            </Text>
+          </Button>
+        </VStack>
+        <VStack marginTop={16}>
+          <Button
+            backgroundColor={'light.100'}
+            w={96}
+            padding={4}
+            onPress={() => console.log('google login')}>
+            <HStack
+              w={96}
+              paddingX={4}
+              alignItems={'center'}
+              justifyContent="space-between">
+              <Icon
+                as={Ionicons}
+                name="logo-google"
+                size="md"
+                color={'black'}
+              />
+              <Text fontSize={'16px'}>구글 아이디로 로그인</Text>
+              <Icon as={Ionicons} name="chevron-forward-outline" size={'md'} />
+            </HStack>
+          </Button>
+        </VStack>
+      </VStack>
+    </Box>
   );
 }
