+++ src/screens/Auth/RegisterScreen.tsx
@@ -1,11 +1,11 @@
-import {Box, Button, NativeBaseProvider, Text, VStack} from 'native-base';
+import {Box, Button, Text, VStack} from 'native-base';
 import React from 'react';
 import Input from '~/components/Input';
 import {useForm, Controller} from 'react-hook-form';
 import * as yup from 'yup';
 import {yupResolver} from '@hookform/resolvers/yup';
 import ErrorMsg from '~/components/ErrorMsg';
-import {signUp} from '~/api/fireAuth';
+import {signUp} from '~/common/api/fireAuth';
 
 interface FormData {
   email: string;
@@ -42,9 +42,31 @@ function RegisterScreen({navigation}: any) {
   } = useForm<FormData>(formOptions);
 
   return (
-    <NativeBaseProvider>
-      <Box h="100%" w="100%" paddingTop={20} backgroundColor="white">
-        <VStack alignItems="center" justifyContent="center" space={12}>
+    <Box h="100%" w="100%" paddingTop={20} backgroundColor="white">
+      <VStack alignItems="center" justifyContent="center" space={12}>
+        <VStack space={3}>
+          <Controller
+            control={control}
+            rules={{
+              required: true,
+            }}
+            render={({field: {onChange, onBlur, value}}) => (
+              <Input
+                onBlur={onBlur}
+                onChangeText={onChange}
+                value={value}
+                type="text"
+                ph="이메일을 입력해주세요"
+              />
+            )}
+            name="email"
+          />
+          {errors.email && (
+            <ErrorMsg>이메일 형식이 올바르지 않습니다!</ErrorMsg>
+          )}
+        </VStack>
+
+        <VStack space={6}>
           <VStack space={3}>
             <Controller
               control={control}
@@ -56,63 +78,17 @@ function RegisterScreen({navigation}: any) {
                   onBlur={onBlur}
                   onChangeText={onChange}
                   value={value}
-                  type="text"
-                  ph="이메일을 입력해주세요"
+                  type="password"
+                  ph="비밀번호를 입력해주세요"
                 />
               )}
-              name="email"
+              name="password"
             />
-            {errors.email && (
-              <ErrorMsg>이메일 형식이 올바르지 않습니다!</ErrorMsg>
+            {errors.password && (
+              <ErrorMsg>비밀번호를 다시 확인해주세요</ErrorMsg>
             )}
           </VStack>
 
-          <VStack space={6}>
-            <VStack space={3}>
-              <Controller
-                control={control}
-                rules={{
-                  required: true,
-                }}
-                render={({field: {onChange, onBlur, value}}) => (
-                  <Input
-                    onBlur={onBlur}
-                    onChangeText={onChange}
-                    value={value}
-                    type="password"
-                    ph="비밀번호를 입력해주세요"
-                  />
-                )}
-                name="password"
-              />
-              {errors.password && (
-                <ErrorMsg>비밀번호를 다시 확인해주세요</ErrorMsg>
-              )}
-            </VStack>
-
-            <VStack space={3}>
-              <Controller
-                control={control}
-                rules={{
-                  required: true,
-                }}
-                render={({field: {onChange, onBlur, value}}) => (
-                  <Input
-                    onBlur={onBlur}
-                    onChangeText={onChange}
-                    value={value}
-                    type="password"
-                    ph="비밀번호를 다시 입력해주세요"
-                  />
-                )}
-                name="checkPassword"
-              />
-              {errors.checkPassword && (
-                <ErrorMsg>비밀번호가 일치하지 않습니다</ErrorMsg>
-              )}
-            </VStack>
-          </VStack>
-
           <VStack space={3}>
             <Controller
               control={control}
@@ -124,27 +100,49 @@ function RegisterScreen({navigation}: any) {
                   onBlur={onBlur}
                   onChangeText={onChange}
                   value={value}
-                  type="text"
-                  ph="닉네임을 입력해주세요"
+                  type="password"
+                  ph="비밀번호를 다시 입력해주세요"
                 />
               )}
-              name="name"
+              name="checkPassword"
             />
-            {errors.name && <ErrorMsg>닉네임을 다시 확인해주세요</ErrorMsg>}
+            {errors.checkPassword && (
+              <ErrorMsg>비밀번호가 일치하지 않습니다</ErrorMsg>
+            )}
           </VStack>
-          <Button
-            marginTop={'10'}
-            colorScheme={'blue'}
-            w={96}
-            padding={'4'}
-            onPress={handleSubmit(onSubmit)}>
-            <Text color={'white'} fontSize={'16px'}>
-              회원가입
-            </Text>
-          </Button>
         </VStack>
-      </Box>
-    </NativeBaseProvider>
+
+        <VStack space={3}>
+          <Controller
+            control={control}
+            rules={{
+              required: true,
+            }}
+            render={({field: {onChange, onBlur, value}}) => (
+              <Input
+                onBlur={onBlur}
+                onChangeText={onChange}
+                value={value}
+                type="text"
+                ph="닉네임을 입력해주세요"
+              />
+            )}
+            name="name"
+          />
+          {errors.name && <ErrorMsg>닉네임을 다시 확인해주세요</ErrorMsg>}
+        </VStack>
+        <Button
+          marginTop={'10'}
+          colorScheme={'blue'}
+          w={96}
+          padding={'4'}
+          onPress={handleSubmit(onSubmit)}>
+          <Text color={'white'} fontSize={'16px'}>
+            회원가입
+          </Text>
+        </Button>
+      </VStack>
+    </Box>
   );
 }
